refactor(frontend): enforce a public-barrel boundary between features (#audit-07)
Resolves the decoupling CRIT (item 7.2.2): no feature reaches into another
feature's internals anymore. Shared leaf bricks (MarkdownViewer, StatusBadge,
bboxScaling) move to shared/; every cross-feature access now goes through the
target feature's public index.ts barrel (@/features/<name>). An ESLint
no-restricted-imports rule enforces the boundary (test files exempt), proven
to block a deep import.

This is the barrel-boundary resolution (Option B) sanctioned by the audit,
not a store dependency inversion — behaviour is unchanged (imports only).
The audit fiche 07 (item 7.2.2) is reworded to encode the convention; the 07
report and summary note the resolution.
