+++ frontend/src/features/admin-config/index.ts
@@ -1,0 +1,2 @@
+export { useAdminConfigStore } from './store'
+export { default as ReasoningConfigSection } from './ui/ReasoningConfigSection.vue'
