+++ frontend/src/features/analysis/graphApi.ts
@@ -1,4 +1,4 @@
-import { apiFetch } from '../../shared/api/http'
+import { apiFetch } from '@/shared/api/http'
 
 /**
  * A single provenance entry for an element — matches Docling's
