+++ frontend/src/features/chunking/api.ts
@@ -1,5 +1,5 @@
 import type { Chunk, ChunkingOptions } from '../../shared/types'
-import { apiFetch } from '../../shared/api/http'
+import { apiFetch } from '@/shared/api/http'
 
 export function rechunkAnalysis(jobId: string, chunkingOptions: ChunkingOptions): Promise<Chunk[]> {
   return apiFetch<Chunk[]>(`/api/analyses/${jobId}/rechunk`, {
