+++ frontend/src/features/document/index.ts
@@ -1,4 +1,5 @@
 export { useDocumentStore } from './store'
+export { getPreviewUrl, rechunkDocument } from './api'
 export { default as DocumentUpload } from './ui/DocumentUpload.vue'
 export { default as DocumentList } from './ui/DocumentList.vue'
 export { default as PagePreview } from './ui/PagePreview.vue'
