+++ frontend/src/features/document/ui/BboxCanvas.vue
@@ -24,7 +24,7 @@
  */
 import { onBeforeUnmount, ref, watch } from 'vue'
 import type { PageElement } from '../../../shared/types'
-import { bboxToRect, computeScale, pointInRect } from '../bboxScaling'
+import { bboxToRect, computeScale, pointInRect } from '@/shared/bboxScaling'
 import { colorFor, UNKNOWN_ELEMENT_COLOR } from '@/shared/elementColors'
 
 const props = defineProps<{
