+++ frontend/src/features/document/ui/ElementProperties.vue
@@ -147,7 +147,7 @@ import type { DocChunk, PageElement } from '../../../shared/types'
 import { useI18n } from '../../../shared/i18n'
 import { bboxToPercent } from '../bboxPercent'
 import { colorFor } from '@/shared/elementColors'
-import MarkdownViewer from '../../analysis/ui/MarkdownViewer.vue'
+import MarkdownViewer from '@/shared/ui/MarkdownViewer.vue'
 import TableModal from './TableModal.vue'
 
 const props = defineProps<{
