+++ frontend/src/pages/RunsPage.vue
@@ -5,7 +5,7 @@
 <script setup lang="ts">
 import { useI18n } from '../shared/i18n'
 
-import ComingSoonShell from '../shared/ui/ComingSoonShell.vue'
+import ComingSoonShell from '@/shared/ui/ComingSoonShell.vue'
 
 const { t } = useI18n()
 </script>
