+++ frontend/src/shared/bboxScaling.test.ts
