+++ frontend/src/shared/ui/StatusBadge.vue
@@ -13,8 +13,8 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 
-import { useI18n } from '../../../shared/i18n'
-import type { DocumentLifecycleState } from '../../../shared/types'
+import { useI18n } from '../i18n'
+import type { DocumentLifecycleState } from '../types'
 
 const props = withDefaults(
   defineProps<{
